requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-auth
google-generativeai
google-adk
firebase-admin